dev-requirements.txt
-r requirements.txt
pdbpp==0.7.2
fabric==1.8.3
path.py==5.1
django-webtest==1.7.7
mock==1.0.1
WebOb==1.4
WebTest==2.0.15
waitress==0.8.9
beautifulsoup4==4.3.2
factory-boy==2.3.1
